refactor: Moved visualization map components to the correct folder
