+++ src/sharedData/visualization/components/VisualizationMapLayer.js
