+++ src/storyMap/components/StoryMap.js
@@ -33,8 +33,8 @@ import './StoryMap.css';
 import logger from 'terraso-client-shared/monitoring/logger';
 
 import Map, { MapContextConsumer } from 'gis/components/Map';
-import MapboxLayer from 'gis/components/MapboxLayer';
-import MapboxRemoteSource from 'gis/components/MapboxRemoteSource';
+import VisualizationMapLayer from 'sharedData/visualization/components/VisualizationMapLayer';
+import VisualizationMapRemoteSource from 'sharedData/visualization/components/VisualizationMapRemoteSource';
 
 import StoryMapOutline from './StoryMapOutline';
 
@@ -554,11 +554,11 @@ const StoryMap = props => {
           {!_.isEmpty(config.dataLayers) &&
             Object.values(config.dataLayers).map(dataLayerConfig => (
               <React.Fragment key={dataLayerConfig.id}>
-                <MapboxRemoteSource
+                <VisualizationMapRemoteSource
                   sourceName={dataLayerConfig.id}
                   visualizationConfig={dataLayerConfig}
                 />
-                <MapboxLayer
+                <VisualizationMapLayer
                   sourceName={dataLayerConfig.id}
                   visualizationConfig={dataLayerConfig}
                   showPopup={false}
